requirements.txt
google-genai>=1.38.0
pyinaturalist>=0.20.2
pandas>=2.3.2
